rework targeting of image tests
We'll now have 3 "categories" of test:

Group 1 are all tests tagged with `smoke`, and not excluded.
These ideally don't take a huge amount of time to run, and should
reliably indicate problems.

Group 2 is all image tests
These take a long time to run, so we'd like scheduling flexibility.

Of these image tests, we want the ability to indicate which ones are
important to alert on, and which ones not. for now, this is done by the
"critical" and "noncritical" test attributes, giving two subgroups.