update build timeout to 40 minutes for baremetal
Baremetal nodes can take quite a long time to deploy, as it requires
2 full reboots of a physical node.

In ironic, `deploy_callback_timeout` defaults to 1800s, after which
deployment will fail with error "timeout waiting ofr call-back".

The end-to-end build timeout is defined in nova, as
`instance_build_timeout`, but is disabled by default.

We set the timeout here to 40 minutes, 2400 seconds, in order to ensure
that we catch the ironic timeout error rather than aborting the
deployment in the test.